% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_clusters}
\alias{add_clusters}
\title{Add cluster information to the Seurat object}
\usage{
add_clusters(scdata, parsed_cellsets, cell_sets)
}
\arguments{
\item{scdata}{Seurat object}

\item{parsed_cellsets}{data.table cellsets object}

\item{cell_sets}{list with the cell_sets meta data (like keys and type)}
}
\value{
Seurat object
}
\description{
This function adds cluster information coming from the parsed cellsets file
to the metadata slot of the Seurat object.
For Louvain/Leiden clusters, it adds a single column with cluster names as values.
For scratchpad clusters, it adds one column for each cluster with TRUE/FALSE values
to indicate if the corresponding cell belongs to that cluster. This is because
a cell can belong to more than one scratchpad cluster.
For ScType annotations, it adds a column for each unique species-tissue annotation.
}
