% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{add_gene_symbols}
\alias{add_gene_symbols}
\title{Add gene symbols to count matrix}
\usage{
add_gene_symbols(scale_data, scdata)
}
\arguments{
\item{scale_data}{data.frame. Count matrix}

\item{scdata}{Seurat object}
}
\value{
count matrix with gene symbols in the new column "original_name"
}
\description{
This function adds gene symbols to a count matrix.
It takes the gene symbols from the misc slot of the Seurat object, which
is created in the pipeline.
}
