% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle_scoring.R
\name{cellCycleScoring}
\alias{cellCycleScoring}
\title{Cell Cycle Scoring Function}
\usage{
cellCycleScoring(req, scdata)
}
\arguments{
\item{req}{The request object contains API URL, experiment ID, authentication JWT.}

\item{scdata}{SeuratObject}
}
\value{
Formatted cell sets based on cell cycle scoring.
}
\description{
This function performs cell cycle scoring on scRNA-seq data. Uses the Seurat method as default.
}
