% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{collapse_genes}
\alias{collapse_genes}
\title{Collapse gene symbols}
\usage{
collapse_genes(scale_data)
}
\arguments{
\item{scale_data}{count matrix}
}
\value{
collapsed count matrix
}
\description{
This function collapses rows of the count matrix by gene symbols.
For each duplicated gene symbol, it takes the row with the highest mean expression.
}
