% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{completeExpression}
\alias{completeExpression}
\title{Adds an empty row for every filtered cell}
\usage{
completeExpression(expression, cell_ids)
}
\arguments{
\item{cell_ids}{integer vector}

\item{rawExpression}{data.table}
}
\value{
complete raw expression data.table
}
\description{
The UI infers cell_id by the index of the cell in the matrix, which means that
filtered cells have to be added back to the table, as empty rows. When converted
to sparse format, they do not take up space.
}
