% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_annot_seurat_object.R
\name{DownloadAnnotSeuratObject}
\alias{DownloadAnnotSeuratObject}
\title{Annotate Seurat object}
\usage{
DownloadAnnotSeuratObject(req, data)
}
\arguments{
\item{req}{Request object that contains the cellsets object and embedding data.}

\item{data}{Seurat object}
}
\value{
Path to the saved Seurat object.
}
\description{
This function adds information about embeddings and clustering to the processed Seurat object.
Embedding are obtained and added to the Seurat object using the assignEmbedding function.
Cellsets object is received from the Python worker and parsed to extract
relevant information about louvain/leiden clusters, custom cellsets,
and cluster annotations. This information is added to the Seurat object
to provide a more comprehensive representation of the data.
}
