% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_errors.R
\name{extractErrorList}
\alias{extractErrorList}
\title{extractErrorList}
\usage{
extractErrorList(error_message)
}
\arguments{
\item{error_message}{generated with generateErrorMessage.}
}
\value{
sepparates an error message into code and user message.
}
\description{
extractErrorList
}
