% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{format_cell_sets_object}
\alias{format_cell_sets_object}
\title{Formats cell sets object for patching through the API}
\usage{
format_cell_sets_object(cell_sets, clustering_method, color_pool)
}
\arguments{
\item{cell_sets}{data.frame with two columns: cluster and cell_ids}

\item{clustering_method}{string Either louvain or leiden.}

\item{color_pool}{character vector of colors in hex}
}
\value{
list
}
\description{
This function is only used to format clustering cellsets. Converting from
data.frame to list and adding slots necessary for the cellsets file.
}
