% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{format_matrix}
\alias{format_matrix}
\title{Format count matrix for ScType}
\usage{
format_matrix(scale_data)
}
\arguments{
\item{scale_data}{data.frame. Count matrix}
}
\value{
formatted count matrix
}
\description{
This function format the count matrix as a matrix with gene symbols as rownames.
}
