% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle_scoring.R
\name{format_phase_cellsets}
\alias{format_phase_cellsets}
\title{Format cell cycle phase cellsets}
\usage{
format_phase_cellsets(cell_sets, color_pool)
}
\arguments{
\item{cell_sets}{data.frame with cell ids and phase}

\item{color_pool}{list}

\item{name}{cell cycle phase name}
}
\value{
list of structured cellsets
}
\description{
Format cell cycle phase cellsets
}
