% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_sctype_cell_sets}
\alias{format_sctype_cell_sets}
\title{Formats ScType cell sets object for patching through the API}
\usage{
format_sctype_cell_sets(data, species, tissue)
}
\arguments{
\item{data}{Seurat object with ScType anntations in the "customclassif" column of the metadata slot}

\item{species}{string. Either "human" or "mouse"}

\item{tissue}{string. Tissue information}
}
\value{
list
}
\description{
This function is used to format ScType cellsets. Converting from
data.frame to list and adding slots necessary for the cellsets file.
}
