% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{formatExpression}
\alias{formatExpression}
\title{formatExpression}
\usage{
formatExpression(data)
}
\arguments{
\item{data}{}
}
\value{
Formats expression values for the UI.
}
\description{
formatExpression
}
