% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_errors.R
\name{formatResponse}
\alias{formatResponse}
\title{format response}
\usage{
formatResponse(data, error)
}
\arguments{
\item{error}{}
}
\value{
formats response for the UI.
}
\description{
format response
}
