% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_capital.R
\name{fund_capital}
\alias{fund_capital}
\title{Fund Capital Allocation}
\usage{
fund_capital(funds_flow, requirements)
}
\arguments{
\item{funds_flow}{A matrix of columns, one for each funding contract and rows, one for each forward period. Each cell with have values for 0 cashflow, -1 unit investment in a contract, and (1+rate)^duration, the unit forward value given a period rate and period duration of the contract.}

\item{requirements}{A vector of forward cashflow requirements of length equal to nrow(funds_flow).}
}
\value{
An unordered list of input funds flow, requirements, solution, and objective value.
}
\description{
This function allows the analyst to solve for the initial seed capital given contract specifications for capital across forward periods needed, all to match cashflow requirements in forward periods. Input to a canonical linear programming model are a cashflow funding matrix with rows of forward claims periods and columns of funding contracts, and a vector of forward claims requirements. The function specifies the objective function and the direction of requirements constraints in each forward period. The solution is the amount of capital needed to match claims in each period of claims experience and contract duration. The objective value is simply the initial capital needed to seed the fund.
}
\examples{
# example code
library(lpSolve)
funds_flow <- matrix (c(-1,   0,    0,  0,
                       1.1,  -1,   -1,  0,
                       0,  1.11,    0, -1,
                       0,     0,    0,  1.18,
                       0,     0, 1.52,  0
                       ), nrow=5, byrow=TRUE)
requirements <- c(7, 5, 25, 11)
funding <- fund_capital(
               funds_flow,
               requirements)
funding$solution
funding$sensitivity
# The shadow prices for each period correctly compute the discount factors for each funding period

}
