% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_flow.R
\name{fund_flow}
\alias{fund_flow}
\title{Funds Flow}
\usage{
fund_flow(contracts, requirements)
}
\arguments{
\item{contracts}{A tibble of three columns: the name of the contracts, the period beginning the contract, the duration of the contract in periods, the period rate of return on the contracts; and rows for each financial contract contracts}

\item{requirements}{A vector of forward period requirements. The first period is the initial capitalization of the fund. Requirements begin with the second period, which is the first forward period after the initial capitalization.}
}
\value{
List of the initial contract specification and a funds flow matrix with nrows = number of contracts, ncols = number of periods indicated by the start and duration of contracts + 1 for the name of the contracts, with column names the list of period labels
}
\description{
We face a need to self-insure workers' compensation claims. This function helps an analyst to build a matrix rows of forward period unit cashflow with columns of funding contracts which specify forward rates, start and finish dates. Returned are the initial contract specification and the cashflow matrix.
}
\examples{
# example code
library( tidyverse )
contracts <- tibble(
name = c( "A", "B", "C", "D" ),
start = c( 1, 2, 2, 3 ),
duration = c( 1, 1, 3, 1),
rate = c( 0.10, 0.11, 0.15, 0.18)
)
requirements <- c( 7, 5, 25, 11 )
fund_flow( contracts, requirements )$funds_flow

}
