% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_helpers.R
\name{generate_request}
\alias{generate_request}
\title{Generates request for worker task}
\usage{
generate_request(task)
}
\arguments{
\item{task}{}
}
\value{

}
\description{
Generates request for worker task
}
