% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_errors.R
\name{generateErrorMessage}
\alias{generateErrorMessage}
\title{generateErrorMessage}
\usage{
generateErrorMessage(code, user_message)
}
\arguments{
\item{code}{error code.}

\item{user_message}{message to display to a user.}
}
\value{

}
\description{
generateErrorMessage
}
