% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{generateGraphData}
\alias{generateGraphData}
\title{Convert Seurat object to Monocle3 cell_data_set object, cluster cells and learn graph}
\usage{
generateGraphData(data)
}
\arguments{
\item{data}{Seurat object}
}
\value{
a cell_data_set object with cluster and graph information stored internally
}
\description{
In order to perform the trajectory analysis with Monocle3, the Seurat object needs to be converted to a Monocle3 cell_data_set object.
This function calculates also embeddings and learn the graph.
}
