% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{generateTrajectoryGraph}
\alias{generateTrajectoryGraph}
\title{Convert Seurat object to Monocle3 cell_data_set object, cluster cells and learn graph}
\usage{
generateTrajectoryGraph(
  embedding_data,
  embedding_settings,
  clustering_settings,
  data
)
}
\arguments{
\item{data}{Seurat object}
}
\value{
a cell_data_set object with cluster and graph information stored internally
}
\description{
In order to perform the trajectory analysis with Monocle3,
the Seurat object needs to be converted to a Monocle3 cell_data_set object.
After conversion, this function also learns the trajectory graph.
}
