% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{get_formatted_data}
\alias{get_formatted_data}
\title{Format count matrix for ScType}
\usage{
get_formatted_data(scdata, active_assay)
}
\arguments{
\item{scdata}{Seurat object}

\item{active_assay}{string. One of: RNA, SCT, integrated. It's to extract scale.data}
}
\value{
formatted scaled count matrix
}
\description{
This function extracts the scaled count matrix from the Seurat object, adds
gene symbols as rownames, and collapses duplicated gene symbols.
The formatted count matrix will be used by ScType to annotate cells.
}
