% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_background_genes.R
\name{getBackgroundExpressedGenes}
\alias{getBackgroundExpressedGenes}
\title{Get background genes for pathway analysis}
\usage{
getBackgroundExpressedGenes(req, data)
}
\arguments{
\item{req}{request parameters}

\item{data}{SeuratObject}
}
\value{
list with gene symbols that are expressed in cells being compared.
}
\description{
Get background genes for pathway analysis
}
