% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_helpers.R
\name{getClusters}
\alias{getClusters}
\title{Compute clusters and return object with clusters}
\usage{
getClusters(type, resolution, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Compute clusters and return object with clusters
}
