% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_information.R
\name{getDoubletScore}
\alias{getDoubletScore}
\title{Extract doublet scores}
\usage{
getDoubletScore(req, data)
}
\arguments{
\item{req}{list request}

\item{data}{seurat object}
}
\description{
Extract doublet scores
}
