% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_cellsets.R
\name{getExpressionCellSet}
\alias{getExpressionCellSet}
\title{Get IDs for a new cell set that matches expression filters.}
\usage{
getExpressionCellSet(req, data)
}
\arguments{
\item{req}{request parameters
e.g. list(
body = list(
list(geneName = 'Gene1', comparisonType = 'greaterThan', thresholdValue = 0.5),
list(geneName = 'Gene2', comparisonType = 'lessThan', thresholdValue = 0.5),
...
)
)}

\item{data}{SeuratObject}
}
\value{
vector of cell ids.
}
\description{
Get IDs for a new cell set that matches expression filters.
}
