% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_information.R
\name{getMitochondrialContent}
\alias{getMitochondrialContent}
\title{Retrieve all the mt-content score for every cell}
\usage{
getMitochondrialContent(req, data)
}
\arguments{
\item{req}{list}

\item{data}{Seurat object}
}
\description{
The MT-content was calculated in GEM2S.
}
