% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_information.R
\name{getNGenes}
\alias{getNGenes}
\title{Retrieve the number of features for every cell}
\usage{
getNGenes(req, data)
}
\arguments{
\item{req}{list}

\item{data}{Seurat object}
}
\description{
Retrieve the number of features for every cell
}
