% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_information.R
\name{getNUmi}
\alias{getNUmi}
\title{Retrieve the number of UMIs for every cell}
\usage{
getNUmi(req, data)
}
\arguments{
\item{req}{list}

\item{data}{Seurat object}
}
\description{
Retrieve the number of UMIs for every cell
}
