% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{getRawExpression}
\alias{getRawExpression}
\title{Extract raw expression values for a list of genes}
\usage{
getRawExpression(data, genes)
}
\arguments{
\item{data}{Seurat object}

\item{genes}{character vector og genes to extract}
}
\value{
data.table of raw expression values
}
\description{
The expression matrix is transposed to accommodate CSC sparse matrix format
used in the UI. Filtered cells are added as empty rows.
}
