% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markers.R
\name{getTopMarkerGenes}
\alias{getTopMarkerGenes}
\title{getTopMarkerGenes}
\usage{
getTopMarkerGenes(
  nFeatures,
  data,
  cellSets,
  aucMin = 0.3,
  pctInMin = 20,
  pctOutMax = 70
)
}
\arguments{
\item{pctOutMax}{}
}
\value{
Returns top marker genes as calculated by presto
}
\description{
getTopMarkerGenes
}
