% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pseudobulk_matrix.R
\name{makePseudobulkMatrix}
\alias{makePseudobulkMatrix}
\title{Make aggregated matrix for pseudo bulk differential expression}
\usage{
makePseudobulkMatrix(scdata)
}
\arguments{
\item{scdata}{a SeuratObject,}
}
\value{
a SeuratObject with counts aggregated by sample for one cluster
}
\description{
The SeuratObject requires a \code{custom} meta.data slot, specifying which cells
belong to the \code{base} and \code{background} groups. It filters out cells with \code{NA} in \code{custom}.
Also, a \code{samples} meta.data slot is required, specifying samples.
}
