% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{quantileTruncate}
\alias{quantileTruncate}
\title{Truncate expression values after quantile threshold}
\usage{
quantileTruncate(x, quantile_threshold)
}
\arguments{
\item{x}{numeric vector of raw expression values}

\item{quantile_threshold}{numeric}
}
\value{
numeric vector of truncated expression values
}
\description{
basically returns x > a => a else x. But takes into account the special case
when more than quantile_threshold percent of the data is 0. extending
the threshold bit by bit until it differs from 0. Therefore preserving some
dynamic range in the adjusted expression values.
}
