% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle_scoring.R
\name{run_cell_cycle_scoring}
\alias{run_cell_cycle_scoring}
\title{run_cell_cycle_scoring}
\usage{
run_cell_cycle_scoring(scdata)
}
\arguments{
\item{scdata}{SeuratObject}
}
\value{
data.frame with cell ids and cycle stage prediction, or undetermined if there is an error.
}
\description{
Uses the Seurat method to calculate cell cycle scores.
}
