% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{run_sctype}
\alias{run_sctype}
\title{Run ScType}
\usage{
run_sctype(data, active_assay, tissue, species)
}
\arguments{
\item{data}{Seurat object}

\item{active_assay}{string. One of: RNA, SCT, integrated. It's to extract scale.data.}

\item{tissue}{string. Tissue information coming from the config. It's used to extract marker genes from the database}

\item{species}{string. Species information. Either "human" or "mouse"}
}
\value{
Seurat object with ScType annotations in the metadata slot
}
\description{
This function runs ScType on the scale.data slot of a Seurat object
producing cell-level annotations that are then merged by cluster.
For more information about ScType see \href{https://github.com/IanevskiAleksandr/sc-type}{this link}.
It adds the resulting annotations as a new column named "customclassif" to the
metadata slot of the Seurat object.
}
