% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runClusters}
\alias{runClusters}
\title{Get Clusters}
\usage{
runClusters(req, data)
}
\arguments{
\item{req}{list with items:
\code{req$body$type} either "louvain" or "leiden"
\code{req$body$config$resolution} integer, range: 0 - 2}

\item{data}{SeuratObject}
}
\value{
data.frame with columns "cluster" and "cell_ids"
}
\description{
Get Clusters
}
