% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{runDE}
\alias{runDE}
\title{Run Differential Expression Task}
\usage{
runDE(req, data)
}
\description{
Runs either a comparison within samples/groups (i.e. marker genes) or
between samples/groups (differential expression analysis). The request post
includes the cell IDs that we need to compare in a cell set named "base"
and "background".
}
