% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{runDotPlot}
\alias{runDotPlot}
\title{Run request for DotPlot}
\usage{
runDotPlot(req, data)
}
\arguments{
\item{req}{{body: {
useMarkerGenes: TRUE/FALSE determines whether to use marker genes or predefined genes
numberOfMarkers: Int. Number of marker genes to use
customGenesList: List of Strings. List of marker genes to use
groupBy: Cellsets to show in the plot. Determines whether to show Louvain/Samples/Custom
filterBy: Cellsets to subset the experiment with.
}
}}

\item{data}{SeuratObject}
}
\value{

}
\description{
Run request for DotPlot
}
