% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{runGenerateTrajectoryGraph}
\alias{runGenerateTrajectoryGraph}
\title{Generate UMAP and nodes coordinates for the initial trajectory analysis plot}
\usage{
runGenerateTrajectoryGraph(data)
}
\arguments{
\item{data}{SeuratObject}
}
\value{
a json with nodes and umap coordinates
}
\description{
Returns a json object with UMAP and nodes coordinates, to be used for the initial plot (embedding + trajectory nodes)
}
\details{
This represents the first step of the trajectory analysis. It allows the creation of the initial trajectory analysis plot, which is composed of an embedding and starting nodes (vertices in the lines).
This plot is used to select the root nodes among the starting nodes. The root nodes will be then used in the following step of the trajectory analysis for pseudotime calculation.
}
