% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{runMarkerHeatmap}
\alias{runMarkerHeatmap}
\title{Generates a marker heatmap}
\usage{
runMarkerHeatmap(req, data)
}
\arguments{
\item{data}{}
}
\value{
list
}
\description{
Generates a marker heatmap
}
