% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_differential_expression.R
\name{runPseudobulkDE}
\alias{runPseudobulkDE}
\title{Run Pseudobulk Differential Expression}
\usage{
runPseudobulkDE(pbulk)
}
\arguments{
\item{pbulk}{a SeuratObject with counts aggregated by sample for one cluster.}
}
\value{
a data.frame of differential expression results.
}
\description{
Run Pseudobulk Differential Expression
}
