% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{runPseudoTimeTask}
\alias{runPseudoTimeTask}
\title{Calculate pseudotime}
\usage{
runPseudoTimeTask(req, data)
}
\arguments{
\item{req}{{body: {
rootNodes: root nodes ids. Determines the root nodes of the trajectory
}
}}

\item{data}{SeuratObject}
}
\value{
a tibble with pseudotime values
}
\description{
Order the cells and generate an array of pseudotime values, based on
the node ids of the root nodes.
}
\details{
Pseudotime is a value that shows the relative progression
of a cell through a path in gene expression space.
These values will be used to plot the cells along a continuous path
that represents the evolution of the process.
This plot represents the final plot of the trajectory analysis, which shows
a map of how cells expression changes, starting from cells with smaller
pseudotime values to cells with larger pseudotimes, along a trajectory.
}
