% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{scaleExpression}
\alias{scaleExpression}
\title{Calculate z-score of gene expression values}
\usage{
scaleExpression(rawExpression)
}
\arguments{
\item{rawExpression}{data.table}
}
\value{
data.table of scaled expression values
}
\description{
Centers values to the mean and scales to the standard deviation.
}
