% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation_sctype.R
\name{ScTypeAnnotate}
\alias{ScTypeAnnotate}
\title{Annotate cells with ScType}
\usage{
ScTypeAnnotate(req, data)
}
\arguments{
\item{req}{list with items:
\code{req$body$cellSets} cellsets object
\code{req$body$species} either "human" or "mouse"
\code{req$body$tissue} one of: Immune system, Pancreas, Liver, Eye, Kidney, Brain, Lung, Adrenal, Heart, Intestine, Muscle, Placenta, Spleen, Stomach, Thymus}

\item{data}{Seurat object}
}
\description{
This function uses \href{https://github.com/IanevskiAleksandr/sc-type}{ScType}
to annotate cells in a Seurat object.
}
\details{
As a preprocessing step, it downloads the cellsets file from s3, parses it
and adds the information as metadata columns in the metadata slot of the Seurat object.
It annotates the cells using ScType, based on Louvain clusters. The annotations
are temporarily added as a new column in the metadata slot of the Seurat object.
Finally, it updates the cellsets object via the API with the cluster-level
annotations generated by ScType.
}
