% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sendCellsetToApi}
\alias{sendCellsetToApi}
\title{Send cell set to the API}
\usage{
sendCellsetToApi(new_cell_set, api_url, experiment_id, cell_set_key, auth_JWT)
}
\arguments{
\item{new_cell_set}{named list of cell sets}

\item{api_url}{string URL of the API}

\item{experiment_id}{string experiment ID}

\item{cell_set_key}{string cell set UUID}

\item{auth_JWT}{string authorization token}
}
\description{
This sends a single, new cell set to the API for patching to the cell sets
file.
}
