% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{sparsify}
\alias{sparsify}
\title{convert data.table to CSC sparse matrix}
\usage{
sparsify(expression)
}
\arguments{
\item{expression}{data.table}
}
\value{
dgCmatrix
}
\description{
NAs are replaced by zeroes by reference. Then coerced to sparse matrix.
}
