% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subsetIds}
\alias{subsetIds}
\title{Subset a Seurat object with the cell ids}
\usage{
subsetIds(scdata, cells_id)
}
\arguments{
\item{scdata}{Seurat object}

\item{cells_id}{Integer vector of cell IDs to keep}
}
\value{
Subsetted Seurat object
}
\description{
Subset a Seurat object with the cell ids
}
