% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{toSparseJson}
\alias{toSparseJson}
\title{extract sparse matrix attributes to mathJS-like sparse matrix format}
\usage{
toSparseJson(matrix)
}
\arguments{
\item{matrix}{}
}
\value{
list with sparse matrix attributes
}
\description{
extract sparse matrix attributes to mathJS-like sparse matrix format
}
