% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{truncateExpression}
\alias{truncateExpression}
\title{Truncates expression values for all genes in data.table}
\usage{
truncateExpression(rawExpression)
}
\arguments{
\item{rawExpression}{data.table}
}
\value{
data.table of truncated gene expression values
}
\description{
Truncates expression values for all genes in data.table
}
