% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{updateCellSetsThroughApi}
\alias{updateCellSetsThroughApi}
\title{Update the cell sets through the API}
\usage{
updateCellSetsThroughApi(
  cell_sets_object,
  api_url,
  experiment_id,
  cell_set_key,
  auth_JWT
)
}
\arguments{
\item{auth_JWT}{}
}
\description{
Used when re-clustering, cell sets are replaced.
}
