% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{completeExpression}
\alias{completeExpression}
\title{Adds an empty row for every filtered cell}
\usage{
completeExpression(expression, cell_ids)
}
\arguments{
\item{expression}{data.table}

\item{cell_ids}{integer vector}
}
\value{
row complete expression data.table
}
\description{
The UI infers cell_id by the index of the cell in the matrix, which means that
filtered cells have to be added back to the table as empty rows. When converted
to sparse format, they do not take up space. We use the max of the cell_ids that
were filtered, which means this table will not contain cells above that. But
it does not change the index of the cells below, and the UI does not care.
}
