% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{formatExpression}
\alias{formatExpression}
\title{Format expression data.table as mathJS json}
\usage{
formatExpression(expression, cell_ids)
}
\arguments{
\item{expression}{data.table with expression values}

\item{cell_ids}{int vector with filtered cell_ids}
}
\value{
list of formatted expression table
}
\description{
Format expression data.table as mathJS json
}
