% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_expression.R
\name{getExpressionValues}
\alias{getExpressionValues}
\title{Get raw, truncated and scaled gene expression values}
\usage{
getExpressionValues(data, genes)
}
\arguments{
\item{data}{Seurat object}

\item{genes}{character vector of genes to extract}
}
\value{
list of expression values
}
\description{
Get raw, truncated and scaled gene expression values
}
