% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_helpers.R
\name{getExpressionValues}
\alias{getExpressionValues}
\title{Returns expression values for selected genes}
\usage{
getExpressionValues(genes, data)
}
\arguments{
\item{genes}{\itemize{
\item Must have names and input(ensmbl ids)
}}

\item{data}{}
}
\value{

}
\description{
Returns expression values for selected genes
}
