% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_matrix.R
\name{GetNormalizedExpression}
\alias{GetNormalizedExpression}
\title{Extract normalized expression matrix}
\usage{
GetNormalizedExpression(req, data)
}
\arguments{
\item{req}{{body: {
filterBy: Cellsets to subset the experiment with
applyFilter: TRUE/FALSE determines whether to subset the data
}
}}

\item{data}{SeuratObject}
}
\description{
This function subsets the Seurat object using the cell IDs received from the
python worker, then extracts and returns the normalized expression matrix.
}
\details{
Data can be subsetted in the python worker according to one of clusters, samples,
and metadata groups, or a combination of them.
}
