% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markers.R
\name{getTopMarkerGenes}
\alias{getTopMarkerGenes}
\title{getTopMarkerGenes}
\usage{
getTopMarkerGenes(
  nFeatures,
  data,
  cellSetsIds,
  aucMin = 0.3,
  pctInMin = 20,
  pctOutMax = 70
)
}
\arguments{
\item{nFeatures}{int number of marker genes to get}

\item{data}{SeuratObject}

\item{aucMin}{min area under the wilcoxon test's ROC for a gene to be
considered a marker}

\item{pctInMin}{min percentage of cells in cellset that have to express a
gene for it to be considered a marker}

\item{pctOutMax}{max percentage of cells outside cellset that can express a
gene for it to be considered a marker}

\item{cellSets}{list of cellsets to split for marker gene selection}
}
\value{
data.frame of top marker genes
}
\description{
Uses presto::wilcoxauc to find the marker genes that distinguish the
cellsets. It then filters the list of genes up to nFeatures using reasonable
defaults.
}
