% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_helpers.R
\name{load_experiment_data}
\alias{load_experiment_data}
\title{Loads Experiment data}
\usage{
load_experiment_data(exp_id = "test")
}
\arguments{
\item{exp_id}{}
}
\value{

}
\description{
Loads Experiment data
}
