% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{runGenerateTrajectoryGraph}
\alias{runGenerateTrajectoryGraph}
\title{Generate node coordinates for the initial trajectory analysis plot}
\usage{
runGenerateTrajectoryGraph(req, data)
}
\arguments{
\item{req}{list of configuration parameters}

\item{data}{SeuratObject}
}
\value{
a list containing nodes coordinates, connected nodes and the node_id
}
\description{
Returns a list containing node coordinates to the python worker
}
\details{
This represents the first step of the trajectory analysis.
It allows the creation of the initial trajectory analysis plot,
which is composed of an embedding and starting nodes (vertices in the lines).
This plot is used to select the root nodes among the starting nodes.
The root nodes will be then used in the following step of the trajectory
analysis for pseudotime calculation.
}
