% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{runMarkerHeatmap}
\alias{runMarkerHeatmap}
\title{Generates a marker heatmap}
\usage{
runMarkerHeatmap(req, data)
}
\arguments{
\item{req}{list with number of genes and cellsets in which to search for markers}

\item{data}{Seurat object}
}
\value{
list
}
\description{
Generates a marker heatmap
}
